requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
qbittorrent-api==2025.7.0
requests==2.31.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
aiofiles==23.2.1
watchdog==3.0.0
mutagen==1.47.0
pyyaml==6.0.3
aiohttp==3.13.2
